requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
loguru==0.7.2
python-multipart==0.0.6
aiohttp==3.9.1
Pillow==10.1.0
tenacity==8.2.3
webdriver-manager==3.9.1
scrapling==1.0.0
playwright==1.41.2
pytest-asyncio==0.23.5